requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
PyYAML==6.0
sounddevice==0.4.6
numpy==1.24.3
transformers==4.30.2
torch==2.0.1
python-dotenv==1.0.0
